+++ CJBCheatsMenu/ModEntry.cs
@@ -2,6 +2,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using CJB.Common;
+using CJB.Common.Integrations;
 using CJBCheatsMenu.Framework;
 using CJBCheatsMenu.Framework.Models;
 using StardewModdingAPI;
@@ -60,6 +61,7 @@ namespace CJBCheatsMenu
             this.Cheats = new PerScreen<CheatManager>(() => new CheatManager(this.Config, this.Helper.GameContent, this.Helper.Reflection, () => this.Locations.Value.Value, () => this.Warps));
 
             // hook events
+            helper.Events.GameLoop.GameLaunched += this.OnGameLaunched;
             helper.Events.Content.AssetRequested += this.OnAssetRequested;
 
             helper.Events.Display.Rendered += this.OnRendered;
@@ -79,6 +81,90 @@ namespace CJBCheatsMenu
         /*********
         ** Private methods
         *********/
+        /// <summary>Raised after the game is launched, right before the first update tick. This happens once per game session (unrelated to loading saves). All mods are loaded and initialised at this point, so this is a good time to set up mod integrations.</summary>
+        /// <param name="sender">The event sender.</param>
+        /// <param name="e">The event arguments.</param>
+        private void OnGameLaunched(object? sender, GameLaunchedEventArgs e)
+        {
+            // get Generic Mod Config Menu's API (if it's installed)
+            var configMenu = this.Helper.ModRegistry.GetApi<IGenericModConfigMenuApi>("spacechase0.GenericModConfigMenu");
+            if (configMenu is null)
+                return;
+
+            // register mod
+            configMenu.Register(
+                mod: this.ModManifest,
+                reset: () => this.Config = new ModConfig(),
+                save: () => this.Helper.WriteConfig(this.Config),
+                titleScreenOnly: true
+            );
+
+            // ---------------------------- //
+            // Main options
+            // ---------------------------- //
+            configMenu.AddSectionTitle(this.ModManifest, I18n.Config_Title_MainOptions);
+
+            configMenu.AddTextOption(
+                mod: this.ModManifest,
+                name: I18n.Config_DefaultTab_Name,
+                tooltip: I18n.Config_DefaultTab_Desc,
+                getValue: () => this.Config.DefaultTab.ToString(),
+                setValue: value => {
+                    if (Enum.TryParse(value, out MenuTab tab))
+                        this.Config.DefaultTab = tab;
+                },
+                allowedValues: ["PlayerAndTools", "FarmAndFishing", "Skills", "Weather", "Relationships", "WarpLocations", "Time", "Advanced", "Controls"]
+            );
+
+            // add grow radious from 1 to 10 integer
+            configMenu.AddNumberOption(
+                mod: this.ModManifest,
+                name: I18n.Controls_GrowRadius,
+                tooltip: I18n.Config_GrowRadius_Desc,
+                getValue: () => this.Config.GrowRadius,
+                setValue: value => this.Config.GrowRadius = (int)value,
+                min: 1,
+                max: 10
+            );
+
+            // ---------------------------- //
+            // Controls
+            // ---------------------------- //
+            configMenu.AddSectionTitle(this.ModManifest, I18n.Controls_Title);
+
+            configMenu.AddKeybindList(
+                mod: this.ModManifest,
+                name: I18n.Controls_OpenMenu,
+                tooltip: I18n.Config_OpenMenu_Desc,
+                getValue: () => this.Config.OpenMenuKey,
+                setValue: value => this.Config.OpenMenuKey = value
+            );
+
+            configMenu.AddKeybindList(
+                mod: this.ModManifest,
+                name: I18n.Controls_FreezeTime,
+                tooltip: I18n.Config_FreezeTime_Desc,
+                getValue: () => this.Config.FreezeTimeKey,
+                setValue: value => this.Config.FreezeTimeKey = value
+            );
+
+            configMenu.AddKeybindList(
+                mod: this.ModManifest,
+                name: I18n.Controls_GrowTree,
+                tooltip: I18n.Config_GrowTree_Desc,
+                getValue: () => this.Config.GrowTreeKey,
+                setValue: value => this.Config.GrowTreeKey = value
+            );
+
+            configMenu.AddKeybindList(
+                mod: this.ModManifest,
+                name: I18n.Controls_GrowCrops,
+                tooltip: I18n.Config_GrowCrops_Desc,
+                getValue: () => this.Config.GrowCropsKey,
+                setValue: value => this.Config.GrowCropsKey = value
+            );
+        }
+
         /// <summary>Raised after the player loads a save slot.</summary>
         /// <param name="sender">The event sender.</param>
         /// <param name="e">The event arguments.</param>